Update lite-paper.md
I'm pretty sure I made my fair share of silly mistakes and typos in this since I ended up having to turn to it later than planned, but wanted to provide a possible idea for an edited Lite Paper. I tried to keep the substance the same except where there was something that was pointed out to me as needing to be changed.